Renamed tools into utils package
