+++ src/main/java/org/oopproject/utils/Config.java
@@ -1,4 +1,4 @@
-package org.oopproject.tools;
+package org.oopproject.utils;
 
 import feign.Feign;
 import feign.gson.GsonDecoder;
