+++ src/main/java/org/oopproject/utils/Genres.java
@@ -1,4 +1,4 @@
-package org.oopproject.tools;
+package org.oopproject.utils;
 
 public enum Genres {
     ANIMATION(16),
